Remove ScrollingSwiftUIViewController to avoid iOS14 freeze
