+++ Wikipedia/Code/ApplePayContentView.swift
@@ -3,54 +3,56 @@ import SwiftUI
 struct ApplePayContentView: View {
     @SwiftUI.State var input: String = ""
     var body: some View {
-        VStack {
-            Group {
-                Text("Testing")
-                Text("Testing")
-                Text("Testing")
-                Text("Testing")
-                Text("Testing")
-                Text("Testing")
-                Text("Testing")
+        ScrollView {
+            VStack {
+                Group {
+                    Text("Testing")
+                    Text("Testing")
+                    Text("Testing")
+                    Text("Testing")
+                    Text("Testing")
+                    Text("Testing")
+                    Text("Testing")
+                }
+                
+                TextField("Enter Text", text: $input, onEditingChanged: { isEditing in
+                    postEditingNotifications(isEditing: isEditing)
+                })
+                .textFieldStyle(.roundedBorder)
+                .padding()
+                
+                Group {
+                    Text("Testing")
+                    Text("Testing")
+                    Text("Testing")
+                    Text("Testing")
+                    Text("Testing")
+                    Text("Testing")
+                    Text("Testing")
+                }
+                
+                Group {
+                    Text("Testing")
+                    Text("Testing")
+                    Text("Testing")
+                    Text("Testing")
+                    Text("Testing")
+                    Text("Testing")
+                    Text("Testing")
+                }
+                
+                Group {
+                    Text("Testing")
+                    Text("Testing")
+                    Text("Testing")
+                    Text("Testing")
+                    Text("Testing")
+                    Text("Testing")
+                    Text("Testing")
+                }
             }
-            
-            TextField("Enter Text", text: $input, onEditingChanged: { isEditing in
-                postEditingNotifications(isEditing: isEditing)
-            })
-            .textFieldStyle(.roundedBorder)
-            .padding()
-            
-            Group {
-                Text("Testing")
-                Text("Testing")
-                Text("Testing")
-                Text("Testing")
-                Text("Testing")
-                Text("Testing")
-                Text("Testing")
-            }
-            
-            Group {
-                Text("Testing")
-                Text("Testing")
-                Text("Testing")
-                Text("Testing")
-                Text("Testing")
-                Text("Testing")
-                Text("Testing")
-            }
-            
-            Group {
-                Text("Testing")
-                Text("Testing")
-                Text("Testing")
-                Text("Testing")
-                Text("Testing")
-                Text("Testing")
-                Text("Testing")
-            }
+            .padding([.top, .bottom], 20)
         }
-        .padding([.top, .bottom], 20)
     }
     
     func postEditingNotifications(isEditing: Bool) {
@@ -60,6 +62,10 @@ struct ApplePayContentView: View {
             NotificationCenter.default.post(name: .swiftUITextfieldDidEndEditing, object: nil)
         }
     }
+}
+
+extension ApplePayContentView: NavBarKeyboardDismissable {
+    
 }
 
 struct ApplePayContentView_Previews: PreviewProvider {
